[bcr-wdc-keys] move generate_blinds from swap-service
